refactor: move the naive UTC helpers to shared_kernel/utils
The path no longer states why they exist, so the module docstring now
carries the warning instead: they belong to the persistence boundary, and
importing them outside a repository means a timestamp lost its timezone
further upstream.
