+++ server/src/shared_kernel/adapters/secondary/sql/__init__.py
@@ -1,6 +1,5 @@
 """Shared SQL infrastructure for repositories."""
 
-from .naive_utc import as_utc, to_naive_utc
 from .sql_base_repository import SQLBaseRepository
 
-__all__ = ["SQLBaseRepository", "as_utc", "to_naive_utc"]
+__all__ = ["SQLBaseRepository"]
